requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==11.3.0
python-dotenv==1.1.1
google-generativeai==0.8.5
firecrawl-py==4.3.6
gunicorn==23.0.0
requests==2.32.3
dj-database-url==2.3.0
psycopg[binary]==3.2.3
whitenoise==6.8.2
# PDF processing libraries (from original medical report analyzer)
PyPDF2>=3.0.0
pdfplumber>=0.9.0
